nyayabot/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==9.5.0
cloudinary==1.36.0
openai==1.3.5
python-decouple==3.8
django-extensions==3.2.3
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0